refactor(config): migrate configuration injections under v1alpha1
