+++ internal/resources/v1alpha1/metrics_exporter.go
@@ -4,7 +4,7 @@ import (
 	"strconv"
 
 	"github.com/robolaunch/robot-operator/internal"
-	"github.com/robolaunch/robot-operator/internal/configure"
+	configure "github.com/robolaunch/robot-operator/internal/configure/v1alpha1"
 	robotv1alpha1 "github.com/robolaunch/robot-operator/pkg/api/roboscale.io/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
